requirements.txt
pdfplumber
python-docx
pytesseract
pillow
pandas
tk